templates/test_template.py
"""Tests for [Problem Name] implementation."""

import pytest
import sys
from pathlib import Path

# Add parent directory to path for imports
sys.path.insert(0, str(Path(__file__).parent.parent))

from problems.[category].[file_name] import Solution


class Test[ProblemName]:
    """Test suite for [Problem Name]."""

    def setup_method(self):
        """Set up test fixtures."""
        self.solution = Solution()

    def test_example_1(self):
        """Test example 1 from problem statement."""
        result = self.solution.solve(None)
        assert result == None

    def test_example_2(self):
        """Test example 2 from problem statement."""
        result = self.solution.solve(None)
        assert result == None

    def test_edge_case_empty(self):
        """Test with empty input."""
        result = self.solution.solve(None)
        assert result == None

    def test_edge_case_single(self):
        """Test with single element."""
        result = self.solution.solve(None)
        assert result == None

    def test_edge_case_large(self):
        """Test with large input."""
        result = self.solution.solve(None)
        assert result == None
